refactor: resolve #12
close: #12
