+++ src/method/controllers/method.controller.ts
@@ -1,12 +1,14 @@
 import { Controller, Get, Param } from '@nestjs/common';
 import { MethodService } from 'src/method/services/method.service';
+import { MethodEntity } from '../entities/method.entity';
 
 @Controller('method')
 export class MethodController {
   constructor(private methodService: MethodService) {}
 
+  // TODO: validate the id parameter
   @Get(':id')
-  getMethodById(@Param('id') id: string) {
+  getMethodById(@Param('id') id: MethodEntity['id']) {
     return this.methodService.getMethodById(id);
   }
 }
