+++ src/method/entities/method.entity.ts
@@ -4,8 +4,8 @@ import { Entity, OneToMany, PrimaryColumn } from 'typeorm';
 
 @Entity('method')
 export class MethodEntity extends AbstractEntity<MethodEntity> {
-  @PrimaryColumn('uuid')
-  id: string;
+  @PrimaryColumn({ type: 'varchar', length: 3 })
+  id: 'NR' | 'PR' | 'PIR';
 
   @OneToMany(() => TicketEntity, (ticket) => ticket.method)
   tickets: TicketEntity[];
