+++ src/plan/plan.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { PlanEntity } from 'src/plan/entities/plan.entity';
-import { PlanHistoryEntity } from 'src/plan/entities/plan_history.entity';
+import { PlanHistoryEntity } from 'src/plan/entities/plan-history.entity';
 import { PlanService } from './services/plan.service';
 import { PlanController } from './controllers/plan.controller';
