+++ src/setting/entities/setting.entity.ts
@@ -1,11 +1,11 @@
 import { AbstractEntity } from 'src/common/entities/abstract.entity';
 import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
 
-@Entity()
-export class SettingEntity extends AbstractEntity<SettingEntity> {
+@Entity({ name: 'settings' })
+export class SettingsEntity extends AbstractEntity<SettingsEntity> {
   @PrimaryGeneratedColumn()
   id: string;
 
   @Column({ type: 'money' })
-  minimumInitMoney: number;
+  minAmountOpenTicket: number;
 }
