+++ src/ticket/entities/ticket-history.entity.ts
@@ -1,22 +1,25 @@
-import { PlanHistoryEntity } from 'src/plan/entities/plan_history.entity';
+import { PlanHistoryEntity } from 'src/plan/entities/plan-history.entity';
 
 import { TicketEntity } from 'src/ticket/entities/ticket.entity';
 import { Entity, Column, ManyToOne, PrimaryColumn } from 'typeorm';
 
-@Entity('ticket_plan_history')
-export class TicketPlanHistoryEntity {
+@Entity('ticket_history')
+export class TicketHistoryEntity {
   @PrimaryColumn({ type: 'date' })
   issueDate: Date;
 
   @Column({ type: 'date' })
   maturityDate: Date;
 
-  @ManyToOne(() => TicketEntity, (ticket) => ticket.ticketPlanHistories)
+  @Column({ type: 'money' })
+  amount: number;
+
+  @ManyToOne(() => TicketEntity, (ticket) => ticket.ticketHistories)
   ticket: TicketEntity;
 
   @ManyToOne(
     () => PlanHistoryEntity,
-    (planHistory) => planHistory.ticketPlanHistories,
+    (planHistory) => planHistory.ticketHistories,
   )
   planHistory: PlanHistoryEntity;
 }
