+++ src/ticket/ticket.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { TicketEntity } from 'src/ticket/entities/ticket.entity';
-import { TicketPlanHistoryEntity } from 'src/ticket/entities/ticket_plan_history.entity';
+import { TicketHistoryEntity } from 'src/ticket/entities/ticket-history.entity';
 import { TicketController } from './controllers/ticket.controller';
 import { TicketService } from './services/ticket.service';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([TicketEntity, TicketPlanHistoryEntity])],
+  imports: [TypeOrmModule.forFeature([TicketEntity, TicketHistoryEntity])],
   controllers: [TicketController],
   providers: [TicketService],
 })
